+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # REPLACE-GLOBAL-WITH-NEXT-VERSION
 - PT-11292 - Adds possibility to enter separate credentials for sandbox mode in first run wizard
+- PT-11498 - Adds PayPal Express Button to QuickView from CMS Extensions plugin
 
 # 1.2.0
 - PT-11233 - Do not show PayPal Express Button on the product detail page, if the product is in clearance